🔮 refactor(examples/with-redux): simplify structure
