+++ examples/with-redux/app/api/identity-count/route.ts
@@ -6,7 +6,7 @@ export async function POST(req: Request, res: Response) {
   const { amount = 1 } = body
 
   // simulate IO latency
-  await new Promise(r => setTimeout(r, 500))
+  await new Promise((r) => setTimeout(r, 500))
 
   return NextResponse.json({ data: amount })
 }
