+++ examples/with-redux/app/components/Counter/Counter.tsx
