+++ examples/with-redux/app/icon.ico
