+++ examples/with-redux/app/layout.tsx
@@ -1,6 +1,6 @@
 /* Components */
 import { Providers } from '@/lib/providers'
-import { Nav } from './components'
+import { Nav } from './components/Nav'
 
 /* Instruments */
 import styles from './styles/layout.module.css'
