+++ examples/with-redux/app/page.tsx
@@ -1,5 +1,5 @@
 /* Components */
-import { Counter } from './components'
+import { Counter } from './components/Counter/Counter'
 
 export default function IndexPage() {
   return <Counter />
