+++ examples/with-redux/lib/redux/index.ts
