+++ examples/with-redux/lib/redux/rootReducer.ts
