+++ examples/with-redux/lib/redux/slices/counterSlice/counterSlice.ts
