+++ examples/with-redux/lib/redux/slices/counterSlice/index.ts
