+++ examples/with-redux/lib/redux/slices/counterSlice/selectors.ts
