+++ examples/with-redux/lib/redux/slices/index.ts
