+++ examples/with-redux/lib/redux/store.ts
