--- examples/with-redux/src/app/components/Counter/index.ts
@@ -1,1 +1,0 @@
-export * from './Counter'
