--- examples/with-redux/src/app/components/index.ts
@@ -1,2 +1,0 @@
-export * from './Counter'
-export * from './Nav'
